% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_QuesClassified.R
\name{answ_category_by_ques_category_via_qalist}
\alias{answ_category_by_ques_category_via_qalist}
\title{答案分类根据问题分类进行处理}
\usage{
answ_category_by_ques_category_via_qalist(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
答案分类根据问题分类进行处理
}
\examples{
answ_category_by_ques_category_via_qalist();
}
