% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_QuesClassified.R
\name{answ_category_query_current}
\alias{answ_category_query_current}
\title{查询最新版本的答案分类}
\usage{
answ_category_query_current(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
查询最新版本的答案分类
}
\examples{
answ_category_query_current();
}
