% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_QuesClassified.R
\name{nsim_answ_segment_current}
\alias{nsim_answ_segment_current}
\title{处理答案分词的当前版本}
\usage{
nsim_answ_segment_current(brand = "JBLH", NSOwn = 1)
}
\arguments{
\item{brand}{品牌}

\item{NSOwn}{是否只保留网商专业词汇报}
}
\value{
返回值
}
\description{
处理答案分词的当前版本
}
\examples{
nsim_answ_segment_current()
}
