% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_segmen_deal.R
\name{nsim_answ_segment_version}
\alias{nsim_answ_segment_version}
\title{处理答案分词问题}
\usage{
nsim_answ_segment_version(brand = "JBLH", version = "V1")
}
\arguments{
\item{brand}{品牌}

\item{version}{版本}
}
\value{
返回值
}
\description{
处理答案分词问题
}
\examples{
nsim_answ_segment_version();
}
