% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_CombineQuesByAnsw.R
\name{nsim_aux_combine_ques}
\alias{nsim_aux_combine_ques}
\title{处理名称}
\usage{
nsim_aux_combine_ques(data)
}
\arguments{
\item{data}{数据}
}
\value{
返回值'
}
\description{
处理名称
}
\examples{
nsim_aux_combine_ques();
}
