% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_word_prepared.R
\name{nsim_dict_category}
\alias{nsim_dict_category}
\title{读取分词的类别}
\usage{
nsim_dict_category(brand = "JBLH", unique = TRUE)
}
\arguments{
\item{brand}{品牌}

\item{unique}{唯一}
}
\value{
返回值
}
\description{
读取分词的类别
}
\examples{
nsim_dict_category();
}
