% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_word_prepared.R
\name{nsim_dict_FWord}
\alias{nsim_dict_FWord}
\title{处理到词汇}
\usage{
nsim_dict_FWord(brand = "JBLH", unique = FALSE)
}
\arguments{
\item{brand}{品牌}

\item{unique}{是否唯一}
}
\value{
返回值
}
\description{
处理到词汇
}
\examples{
 nsim_dict_FWord()
}
