% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_CombineQuesByAnsw.R
\name{nsim_find_standardQues}
\alias{nsim_find_standardQues}
\title{查询标准问题}
\usage{
nsim_find_standardQues(data)
}
\arguments{
\item{data}{数据}
}
\value{
返回值
}
\description{
查询标准问题
}
\examples{
nsim_find_standardQues()
}
