% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_qalist_prepared.R
\name{nsim_item_answ_count}
\alias{nsim_item_answ_count}
\title{处理答案的在原始语料的词频率}
\usage{
nsim_item_answ_count(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
处理答案的在原始语料的词频率
}
\examples{
nsim_item_answ_count();
}
