% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_qalist_prepared.R
\name{nsim_item_answ_count_currentVersion}
\alias{nsim_item_answ_count_currentVersion}
\title{获取数据的当前版本}
\usage{
nsim_item_answ_count_currentVersion(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
获取数据的当前版本
}
\examples{
nsim_item_answ_count_currentVersion();
}
