% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_qalist_prepared.R
\name{nsim_item_answ_currentVersion}
\alias{nsim_item_answ_currentVersion}
\title{处理当前版本经过nscl与nsbl后的版本内容}
\usage{
nsim_item_answ_currentVersion(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
处理当前版本经过nscl与nsbl后的版本内容
}
\examples{
nsim_item_answ_currentVersion();
}
