% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_qalist_prepared.R
\name{nsim_item_ques_count}
\alias{nsim_item_ques_count}
\title{处理问题的数据}
\usage{
nsim_item_ques_count(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
处理问题的数据
}
\examples{
nsim_item_ques_count();
}
