% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_qalist_prepared.R
\name{nsim_item_ques_count_currentVersion}
\alias{nsim_item_ques_count_currentVersion}
\title{处理当前版本的问题频率}
\usage{
nsim_item_ques_count_currentVersion(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
处理当前版本的问题频率
}
\examples{
nsim_item_ques_count_currentVersion();
}
