% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_qalist_prepared.R
\name{nsim_item_ques_currentVersion}
\alias{nsim_item_ques_currentVersion}
\title{查询问题经过nsbl与nscl后的当前版本}
\usage{
nsim_item_ques_currentVersion(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
查询问题经过nsbl与nscl后的当前版本
}
\examples{
nsim_item_ques_currentVersion();
}
