% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_segmen_deal.R
\name{nsim_nsdict_version}
\alias{nsim_nsdict_version}
\title{处理网商分词情况}
\usage{
nsim_nsdict_version(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
处理网商分词情况
}
\examples{
nsim_nsdict_version();
}
