% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_qalist_prepared.R
\name{nsim_qalist_fullTxt_current}
\alias{nsim_qalist_fullTxt_current}
\title{提前当前版本}
\usage{
nsim_qalist_fullTxt_current(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
提前当前版本
}
\examples{
nsim_qalist_fullTxt_current();
}
