% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_QuesClassified.R
\name{nsim_ques_segment_current}
\alias{nsim_ques_segment_current}
\title{获取当前版本}
\usage{
nsim_ques_segment_current(brand = "JBLH", NSOwn = 1)
}
\arguments{
\item{brand}{品牌}

\item{NSOwn}{是否当保存网商专业词汇}
}
\value{
返回值
}
\description{
获取当前版本
}
\examples{
nsim_ques_segment_current();
}
