% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_segmen_deal.R
\name{nsim_ques_segment_version}
\alias{nsim_ques_segment_version}
\title{处理问题分类}
\usage{
nsim_ques_segment_version(brand = "JBLH", version = "V1")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
处理问题分类
}
\examples{
nsim_ques_seg();
}
