% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_segmen_deal.R
\name{nsim_segment_version}
\alias{nsim_segment_version}
\title{处理分词通用处理方式}
\usage{
nsim_segment_version(id, txt, brand, version, table_name)
}
\arguments{
\item{id}{ID}

\item{txt}{内容将被进行分词处理}

\item{brand}{品牌}

\item{version}{版本}

\item{table_name}{数据库名称}
}
\value{
返回值
}
\description{
处理分词通用处理方式
}
\examples{
nsim_segment_version();
}
