% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_QuesClassified.R
\name{ques_category_byDictSegment_current}
\alias{ques_category_byDictSegment_current}
\title{返回最新的问题分类数据
 sql:vw_ques_category_byDictSegment_current}
\usage{
ques_category_byDictSegment_current(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
返回最新的问题分类数据
 sql:vw_ques_category_byDictSegment_current
}
\examples{
ques_category_byDictSegment_current
}
