% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_QuesClassified.R
\name{ques_category_byDictSegment_underline}
\alias{ques_category_byDictSegment_underline}
\title{针对问题进行处理}
\usage{
ques_category_byDictSegment_underline(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
针对问题进行处理
}
\examples{
ques_category_byDictSegment_underline();
}
