% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_QuesClassified.R
\name{ques_category_query_current}
\alias{ques_category_query_current}
\title{查询设置问题的分类当前分类，经过分词处理后}
\usage{
ques_category_query_current(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
查询设置问题的分类当前分类，经过分词处理后
}
\examples{
ques_category_query_current();
}
