% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_CombineQuesByAnsw.R
\name{ques_combine_ByAnsw}
\alias{ques_combine_ByAnsw}
\title{提供标准问题,进行版本化存储}
\usage{
ques_combine_ByAnsw(brand = "JBLH", version = "V1")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
提供标准问题,进行版本化存储
}
\examples{
ques_getStandarOne();


}
