% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_CombineQuesByAnsw.R
\name{ques_combine_ByAnsw_query_like}
\alias{ques_combine_ByAnsw_query_like}
\title{处理问题合并，生成类似问题}
\usage{
ques_combine_ByAnsw_query_like(brand = "JBLH")
}
\arguments{
\item{brand}{品牌}
}
\value{
返回值
}
\description{
处理问题合并，生成类似问题
}
\examples{
ques_combine_ByAnsw_query_like();
}
